+++ CAESAR/src/scheduler/Scheduler.java
@@ -7,7 +7,6 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicLong;
 import event.*;
 import run.*;
 import transaction.*;
@@ -38,7 +37,7 @@ public abstract class Scheduler implements Runnable {
 	
 	long startOfSimulation;
 	boolean optimized;
-	AtomicLong total_exe_time;
+	AtomicInteger total_exe_time;
 	
 	AtomicBoolean accidentWarningsFailed;
 	AtomicBoolean tollNotificationsFailed;
@@ -46,7 +45,7 @@ public abstract class Scheduler implements Runnable {
 	Scheduler (int max_x, boolean both_d, int lastS,
 			HashMap<RunID,Run> rs, EventQueues evqueues, ExecutorService exe, 
 			AtomicInteger distrProgr, HashMap<Double,Double> distrFinishT, HashMap<Double,Double> schedStartT, CountDownLatch trans_numb, CountDownLatch d,  
-			long start, boolean opt, AtomicLong total_exe) {
+			long start, boolean opt, AtomicInteger total_exe) {
 		
 		max_xway = max_x;
 		both_dirs = both_d;
