+++ CAESAR/src/scheduler/TimeDrivenScheduler.java
@@ -4,7 +4,6 @@ import java.util.HashMap;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicLong;
 import run.*;
 import distributor.*;
 
@@ -19,7 +18,7 @@ public class TimeDrivenScheduler extends Scheduler implements Runnable {
 	public TimeDrivenScheduler (int max_xway, boolean both_dirs, int lastSec,
 			HashMap<RunID,Run> runs, EventQueues eventqueues, ExecutorService executor, 
 			AtomicInteger distrProgr, HashMap<Double,Double> distrFinishT, HashMap<Double,Double> schedStartT, CountDownLatch trans_numb, CountDownLatch done,  
-			long start, boolean opt, AtomicLong total_exe_time) {	
+			long start, boolean opt, AtomicInteger total_exe_time) {	
 		
 		super(max_xway, both_dirs, lastSec, runs, eventqueues, executor, distrProgr, distrFinishT, schedStartT, trans_numb, done, start, opt, total_exe_time);			
 	}
