+++ CAESAR/src/transaction/Transaction.java
@@ -3,10 +3,9 @@ package transaction;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 import run.*;
 import event.*;
-
 
 /** 
  * A transaction has a sequence of events with the same time stamp and belonging to the same run,
@@ -19,11 +18,11 @@ public abstract class Transaction implements Runnable {
 	HashMap<RunID,Run> runs;
 	
 	long startOfSimulation;
-	AtomicLong total_exe_time;
+	AtomicInteger total_exe_time;
 	
 	public CountDownLatch transaction_number;
 				
-	public Transaction (ArrayList<PositionReport> eventList, HashMap<RunID,Run> rs, long start, AtomicLong tet) {
+	public Transaction (ArrayList<PositionReport> eventList, HashMap<RunID,Run> rs, long start, AtomicInteger tet) {
 		
 		events = eventList;		
 		runs = rs;
